Extrac nav visibility functions into separate file.
